requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
PyYAML>=6.0